+++ relayer-cli/src/config.rs
@@ -77,7 +77,7 @@ pub fn validate_config(config: &Config) -> Result<(), Error> {
 /// b) greater or equal to 1/3
 /// c) strictly less than 1
 fn validate_trust_threshold(id: &ChainId, trust_threshold: TrustThreshold) -> Result<(), Error> {
-    if trust_threshold.denominator == 0 {
+    if trust_threshold.denominator() == 0 {
         return Err(Error::invalid_trust_threshold(
             trust_threshold,
             id.clone(),
@@ -85,7 +85,7 @@ fn validate_trust_threshold(id: &ChainId, trust_threshold: TrustThreshold) -> Re
         ));
     }
 
-    if trust_threshold.numerator * 3 < trust_threshold.denominator {
+    if trust_threshold.numerator() * 3 < trust_threshold.denominator() {
         return Err(Error::invalid_trust_threshold(
             trust_threshold,
             id.clone(),
@@ -93,7 +93,7 @@ fn validate_trust_threshold(id: &ChainId, trust_threshold: TrustThreshold) -> Re
         ));
     }
 
-    if trust_threshold.numerator >= trust_threshold.denominator {
+    if trust_threshold.numerator() >= trust_threshold.denominator() {
         return Err(Error::invalid_trust_threshold(
             trust_threshold,
             id.clone(),