+++ relayer/src/worker/client.rs
@@ -122,8 +122,11 @@ impl ClientWorker {
 
                 Next::Continue
             }
-            WorkerCmd::Shutdown => Next::Abort,
+
             WorkerCmd::NewBlock { .. } => Next::Continue,
+            WorkerCmd::ClearPendingPackets => Next::Continue,
+
+            WorkerCmd::Shutdown => Next::Abort,
         }
     }